audit(2.9.7): u128 migration complete + G-01a/G-01g/H-02a gates
F-010: 18 src/test files updated to u128, burn.rs add_fees_burned u64 to u128
G-01g: test_reorg_6_blocks_orphan_scenario written and passing
H-02a: MAINNET_CONSTITUTION.md FROZEN SHA-256 c76aa00224a37ba52950f7ce9e2f72cfc87aa84f5599d3ea28e57538441a8d97
audit: 460 tests 0 failures, 95/102 items complete

+++ L1/core/src/blockchain/burn.rs
@@ -127,9 +127,9 @@ pub struct BuybackStats {
     /// Total ZION sent to DAO (atomic units)
     pub total_zion_creators_rent_atomic: u64,
     /// Total ZION burned via L1 fee burning (tracked separately)
-    pub total_fees_burned_atomic: u64,
+    pub total_fees_burned_atomic: u128,
     /// Combined burn total (only fees — no BTC revenue burn)
-    pub combined_burn_atomic: u64,
+    pub combined_burn_atomic: u128,
     /// Number of revenue events
     pub buyback_count: u64,
     /// Timestamp of last event
@@ -158,7 +158,7 @@ pub struct BuybackStats {
 #[derive(Clone)]
 pub struct BuybackTracker {
     events: Arc<RwLock<Vec<BuybackEvent>>>,
-    fees_burned: Arc<RwLock<u64>>,
+    fees_burned: Arc<RwLock<u128>>,
     ledger_path: String,
 }
 
@@ -169,7 +169,7 @@ impl BuybackTracker {
         let events = Self::load_from_disk(&ledger_path);
         Self {
             events: Arc::new(RwLock::new(events)),
-            fees_burned: Arc::new(RwLock::new(0)),
+            fees_burned: Arc::new(RwLock::new(0u128)),
             ledger_path,
         }
     }
@@ -178,7 +178,7 @@ impl BuybackTracker {
     pub fn in_memory() -> Self {
         Self {
             events: Arc::new(RwLock::new(Vec::new())),
-            fees_burned: Arc::new(RwLock::new(0)),
+            fees_burned: Arc::new(RwLock::new(0u128)),
             ledger_path: String::new(),
         }
     }
@@ -223,7 +223,7 @@ impl BuybackTracker {
 
     /// Record fees burned in a block (called by state.process_block).
     /// Note: This is L1 fee burning, NOT BTC revenue burn.
-    pub fn add_fees_burned(&self, amount_atomic: u64) {
+    pub fn add_fees_burned(&self, amount_atomic: u128) {
         let mut fees = self.fees_burned.write().unwrap();
         *fees = fees.saturating_add(amount_atomic);
     }
